requirements.txt
aiogram
python-dotenv
Pillow
pandas
openpyxl
python-docx
pytesseract
xlrd